*** empty log message ***

+++ libcob/fileio.h
@@ -58,7 +58,7 @@ struct cob_file_desc {
   char access_mode;		/* ACCESS MODE */
   char open_mode;		/* OPEN MODE */
   char *file_status;		/* FILE STATUS */
-  int record_size;		/* record size */
+  size_t record_size;		/* record size */
   unsigned char *record_data;	/* record data address */
   int record_min, record_max;	/* record min/max size */
   struct cob_field record_depending; /* record size depending on */
